moves the tests near the implementation
